*** empty log message ***
git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@1711 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/src/iup_progressbar.c
@@ -32,7 +32,7 @@ void iProgressBarCropValue(Ihandle* ih)
 
 char* iProgressBarGetValueAttrib(Ihandle* ih)
 {
-  return iupStrReturnFloat(ih->data->value);
+  return iupStrReturnFloat((float)ih->data->value);
 }
 
 char* iProgressBarGetDashedAttrib(Ihandle* ih)
